Node.js
Starting to experiment with Node js for use in the next stage of the project.

Have presently replaced the python-server with an equivalent node-server. Also have the .bat file combine the various .js scripts I've been working with in order to minimise number of GET requests sent to our server.

Will now tidy the working directory, think about moving some of our functionality to the server-side (I'd particularly like to move reading of external job pages to the server), and how to use node and SQLite together.